fix: Handle bigger int values on movie details

+++ app/src/main/java/pl/dkd/movieapp/data/movies/model/MovieDetailsDTO.kt
@@ -4,6 +4,6 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class MovieDetailsDTO(
-    val budget: Int,
-    val revenue: Int
+    val budget: Long,
+    val revenue: Long
 )
